requirements.txt
Flask==2.3.3
requests==2.31.0
price-parser==1.0.3
pytz==2023.3
python-dotenv==1.0.0